+++ source/Public/Invoke-SqlDscQuery.ps1
@@ -8,8 +8,25 @@
     .PARAMETER ServerObject
         Specifies current server connection object.
 
+    .PARAMETER ServerName
+        Specifies the server name where the instance exist.
+
+    .PARAMETER InstanceName
+       Specifies the instance name on which to execute the T-SQL query.
+
+    .PARAMETER Credential
+        Specifies the credentials to use to impersonate a user when connecting.
+        If this is not provided then the current user will be used to connect
+        to the SQL Server Database Engine instance.
+
+    .PARAMETER LoginType
+        Specifies which type of credentials are specified. The valid types are
+        Integrated, WindowsUser, and SqlLogin. If WindowsUser or SqlLogin are
+        specified then the Credential needs to be specified as well. Defaults
+        to `Integrated`.
+
     .PARAMETER DatabaseName
-        Specify the name of the database to execute the query on.
+        Specifies the name of the database to execute the T-SQL query in.
 
     .PARAMETER Query
         The query string to execute.
@@ -30,43 +47,71 @@
         outputs none.
 
     .EXAMPLE
-        $serverInstance = Connect-SqlDscDatabaseEngine
-        Invoke-SqlDscQuery -ServerObject $serverInstance -Database master `
+        $serverObject = Connect-SqlDscDatabaseEngine
+        Invoke-SqlDscQuery -ServerObject $serverObject -DatabaseName 'master' `
             -Query 'SELECT name FROM sys.databases' -PassThru
 
-        Runs the query and returns all the database names in the instance.
+        Connects to the default instance and then runs a query to return all the
+        database names in the instance.
 
     .EXAMPLE
-        $serverInstance = Connect-SqlDscDatabaseEngine
-        Invoke-SqlDscQuery -ServerObject $serverInstance -Database master `
+        $serverObject = Connect-SqlDscDatabaseEngine
+        $serverObject | Invoke-SqlDscQuery -DatabaseName 'master' `
             -Query 'RESTORE DATABASE [NorthWinds] WITH RECOVERY'
 
-        Runs the query to restores the database NorthWinds.
+        Connects to the default instance and then runs the query to restore the
+        database NorthWinds.
 
     .EXAMPLE
-        $serverInstance = Connect-SqlDscDatabaseEngine
-        Invoke-SqlDscQuery -ServerObject $serverInstance -Database master `
+        $serverObject = Connect-SqlDscDatabaseEngine
+        Invoke-SqlDscQuery -ServerObject $serverObject -DatabaseName 'master' `
             -Query "select * from MyTable where password = 'PlaceholderPa\ssw0rd1' and password = 'placeholder secret passphrase'" `
-            -PassThru -RedactText @('PlaceholderPa\sSw0rd1','Placeholder Secret PassPhrase') `
-            -Verbose
+            -RedactText @('PlaceholderPa\sSw0rd1','Placeholder Secret PassPhrase') `
+            -PassThru -Verbose
 
         Shows how to redact sensitive information in the query when the query string
-        is output as verbose information when the parameter Verbose is used.
+        is output as verbose information when the parameter Verbose is used. For it
+        to work the sensitiv information must be known and passed into the parameter
+        RedactText. If any single character is wrong the sensitiv information will
+        not be redacted. The redaction is case-insensitive.
 
-    .NOTES
-        This is a wrapper for private function Invoke-Query, until it move into
-        this public function.
+    .EXAMPLE
+        Invoke-SqlDscQuery -ServerName Server1 -InstanceName MSSQLSERVER -DatabaseName 'master' `
+            -Query 'SELECT name FROM sys.databases' -PassThru
+
+        Connects to the default instance and then runs a query to return all the
+        database names in the instance.
 #>
 function Invoke-SqlDscQuery
 {
     [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute('UseSyntacticallyCorrectExamples', '', Justification = 'Because the rule does not yet support parsing the code when a parameter type is not available. The ScriptAnalyzer rule UseSyntacticallyCorrectExamples will always error in the editor due to https://github.com/indented-automation/Indented.ScriptAnalyzerRules/issues/8.')]
     [OutputType([System.Data.DataSet])]
-    [CmdletBinding()]
+    [CmdletBinding(DefaultParameterSetName = 'ByServerName', SupportsShouldProcess = $true, ConfirmImpact = 'High')]
     param
     (
-        [Parameter(Mandatory = $true, ValueFromPipeline = $true)]
+        [Parameter(ParameterSetName = 'ByServerObject', Mandatory = $true, ValueFromPipeline = $true)]
         [Microsoft.SqlServer.Management.Smo.Server]
         $ServerObject,
+
+        [Parameter(ParameterSetName = 'ByServerName')]
+        [ValidateNotNullOrEmpty()]
+        [System.String]
+        $ServerName = (Get-ComputerName),
+
+        [Parameter(ParameterSetName = 'ByServerName')]
+        [System.String]
+        $InstanceName = 'MSSQLSERVER',
+
+        [Parameter(ParameterSetName = 'ByServerName')]
+        [Alias('SetupCredential')]
+        [Alias('DatabaseCredential')]
+        [System.Management.Automation.PSCredential]
+        $Credential,
+
+        [Parameter(ParameterSetName = 'ByServerName')]
+        [ValidateSet('Integrated', 'WindowsUser', 'SqlLogin')]
+        [System.String]
+        $LoginType = 'Integrated',
 
         [Parameter(Mandatory = $true)]
         [System.String]
@@ -77,6 +122,7 @@ function Invoke-SqlDscQuery
         $Query,
 
         [Parameter()]
+        [Alias('WithResults')]
         [Switch]
         $PassThru,
 
@@ -87,32 +133,127 @@ function Invoke-SqlDscQuery
 
         [Parameter()]
         [System.String[]]
-        $RedactText
+        $RedactText,
+
+        [Parameter()]
+        [System.Management.Automation.SwitchParameter]
+        $Force
     )
+
+    begin
+    {
+        if ($Force.IsPresent)
+        {
+            $ConfirmPreference = 'None'
+        }
+
+        if ($PSCmdlet.ParameterSetName -eq 'ByServerName')
+        {
+            $connectSqlDscDatabaseEngineParameters = @{
+                ServerName       = $ServerName
+                InstanceName     = $InstanceName
+                StatementTimeout = $StatementTimeout
+                ErrorAction      = 'Stop'
+                Verbose          = $VerbosePreference
+            }
+
+            if ($LoginType -ne 'Integrated')
+            {
+                $connectSqlDscDatabaseEngineParameters['LoginType'] = $LoginType
+            }
+
+            if ($PSBoundParameters.ContainsKey('Credential'))
+            {
+                $connectSqlDscDatabaseEngineParameters.Credential = $Credential
+            }
+
+            $ServerObject = Connect-SqlDscDatabaseEngine @connectSqlDscDatabaseEngineParameters
+        }
+    }
 
     process
     {
-        $invokeQueryParameters = @{
-            SqlServerObject  = $ServerObject
-            Database         = $DatabaseName
-            Query            = $Query
-        }
+        $redactedQuery = $Query
 
-        if ($PSBoundParameters.ContainsKey('PassThru'))
+        foreach ($redactString in $RedactText)
         {
-            $invokeQueryParameters.WithResults = $PassThru
+            <#
+                Escaping the string to handle strings which could look like
+                regular expressions, like passwords.
+            #>
+            $escapedRedactedString = [System.Text.RegularExpressions.Regex]::Escape($redactString)
+
+            $redactedQuery = $redactedQuery -ireplace $escapedRedactedString, '*******' # cSpell: ignore ireplace
         }
 
-        if ($PSBoundParameters.ContainsKey('StatementTimeout'))
+        $result = $null
+
+        $verboseDescriptionMessage = $script:localizedData.Query_Invoke_ShouldProcessVerboseDescription -f $InstanceName
+        $verboseWarningMessage = $script:localizedData.Query_Invoke_ShouldProcessVerboseWarning -f $InstanceName
+        $captionMessage = $script:localizedData.Query_Invoke_ShouldProcessCaption
+
+        if ($PSCmdlet.ShouldProcess($verboseDescriptionMessage, $verboseWarningMessage, $captionMessage))
         {
-            $invokeQueryParameters.StatementTimeout = $StatementTimeout
-        }
+            $previousStatementTimeout = $null
 
-        if ($PSBoundParameters.ContainsKey('RedactText'))
+            if ($PSCmdlet.ParameterSetName -eq 'ByServerObject')
+            {
+                if ($PSBoundParameters.ContainsKey('StatementTimeout'))
+                {
+                    # Make sure we can return the StatementTimeout before exiting.
+                    $previousStatementTimeout = $ServerObject.ConnectionContext.StatementTimeout
+
+                    $ServerObject.ConnectionContext.StatementTimeout = $StatementTimeout
+                }
+            }
+
+            try
+            {
+                if ($PassThru)
+                {
+                    Write-Verbose -Message (
+                        $script:localizedData.Query_Invoke_ExecuteQueryWithResults -f $redactedQuery
+                    ) -Verbose
+
+                    $result = $ServerObject.Databases[$DatabaseName].ExecuteWithResults($Query)
+
+                    return $result
+                }
+                else
+                {
+                    Write-Verbose -Message (
+                        $script:localizedData.Query_Invoke_ExecuteNonQuery -f $redactedQuery
+                    ) -Verbose
+
+                    $null = $ServerObject.Databases[$DatabaseName].ExecuteNonQuery($Query)
+                }
+            }
+            catch
+            {
+                $writeErrorParameters = @{
+                    Message = $_.Exception.ToString()
+                    Category = 'InvalidOperation'
+                    ErrorId = 'ISDQ0001' # cSpell: disable-line
+                    TargetObject = $DatabaseName
+                }
+
+                Write-Error @writeErrorParameters
+            }
+            finally
+            {
+                if ($previousStatementTimeout)
+                {
+                    $ServerObject.ConnectionContext.StatementTimeout = $previousStatementTimeout
+                }
+            }
+        }
+    }
+
+    end
+    {
+        if ($PSCmdlet.ParameterSetName -eq 'ByServerName')
         {
-            $invokeQueryParameters.RedactText = $RedactText
+            $ServerObject | Disconnect-SqlDscDatabaseEngine -Force -Verbose:$VerbosePreference
         }
-
-        return (Invoke-Query @invokeQueryParameters)
     }
 }